+++ src/elements/conditions/assets/AssetCondition.php
@@ -19,12 +19,12 @@ class AssetCondition extends ElementCondition
     {
         return array_merge(parent::conditionRuleTypes(), [
             DateModifiedConditionRule::class,
-            EditableConditionRule::class,
             FileSizeConditionRule::class,
             FileTypeConditionRule::class,
             FilenameConditionRule::class,
             HasAltConditionRule::class,
             HeightConditionRule::class,
+            SavableConditionRule::class,
             UploaderConditionRule::class,
             ViewableConditionRule::class,
             VolumeConditionRule::class,
