+++ src/elements/conditions/assets/SavableConditionRule.php
@@ -1,0 +1,53 @@
+<?php
+
+namespace craft\elements\conditions\assets;
+
+use Craft;
+use craft\base\conditions\BaseLightswitchConditionRule;
+use craft\base\ElementInterface;
+use craft\elements\conditions\ElementConditionRuleInterface;
+use craft\elements\db\AssetQuery;
+use craft\elements\db\ElementQueryInterface;
+
+/**
+ * Asset savable condition rule.
+ *
+ * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @since 4.4.0
+ */
+class SavableConditionRule extends BaseLightswitchConditionRule implements ElementConditionRuleInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public function getLabel(): string
+    {
+        return Craft::t('app', 'Savable');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getExclusiveQueryParams(): array
+    {
+        return ['savable'];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function modifyQuery(ElementQueryInterface $query): void
+    {
+        /** @var AssetQuery $query */
+        $query->savable($this->value);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function matchElement(ElementInterface $element): bool
+    {
+        $savable = Craft::$app->getElements()->canSave($element);
+        return $savable === $this->value;
+    }
+}
