+++ src/elements/conditions/entries/EditableConditionRule.php
@@ -2,53 +2,14 @@
 
 namespace craft\elements\conditions\entries;
 
-use Craft;
-use craft\base\conditions\BaseLightswitchConditionRule;
-use craft\base\ElementInterface;
-use craft\elements\conditions\ElementConditionRuleInterface;
-use craft\elements\db\ElementQueryInterface;
-use craft\elements\db\EntryQuery;
-use craft\elements\Entry;
-
 /**
- * Entry editable condition rule.
- *
- * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
- * @since 4.3.0
+ * @deprecated in 4.4.0. [[SavableConditionRule]] should be used instead.
+ * @phpstan-ignore-next-line
  */
-class EditableConditionRule extends BaseLightswitchConditionRule implements ElementConditionRuleInterface
-{
-    /**
-     * @inheritdoc
-     */
-    public function getLabel(): string
+if (false) {
+    class EditableConditionRule
     {
-        return Craft::t('app', 'Editable');
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function getExclusiveQueryParams(): array
-    {
-        return ['savable'];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function modifyQuery(ElementQueryInterface $query): void
-    {
-        /** @var EntryQuery $query */
-        $query->savable($this->value);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function matchElement(ElementInterface $element): bool
-    {
-        $savable = Craft::$app->getElements()->canSave($element);
-        return $savable === $this->value;
     }
 }
+
+class_alias(SavableConditionRule::class, EditableConditionRule::class);
