+++ src/translations/en/app.php
@@ -571,7 +571,6 @@ return [
     'Edit “{entryType}”' => 'Edit “{entryType}”',
     'Edit “{title}”' => 'Edit “{title}”',
     'Edit' => 'Edit',
-    'Editable' => 'Editable',
     'Edited' => 'Edited',
     'Eircode' => 'Eircode',
     'Element' => 'Element',
@@ -1224,6 +1223,7 @@ return [
     'SVG file recommended. The logo will be displayed at {size} wide.' => 'SVG file recommended. The logo will be displayed at {size} wide.',
     'Same as asset filesystem' => 'Same as asset filesystem',
     'Same as language' => 'Same as language',
+    'Savable' => 'Savable',
     'Save and add another' => 'Save and add another',
     'Save and continue editing' => 'Save and continue editing',
     'Save and edit entry types' => 'Save and edit entry types',
